Fix high memory usage during thread building
If only messages with threading info are loaded we can significantly
reduce the amount of memory consumed to build threads. This is because
many messages are not part of a thread. And if they are, the references
and in-reply-to headers are enough information to know that there was a
previous message. So we build the threads out of existance of the thread
roots. As long as the threading result is the same, this doesn't matter.

Signed-off-by: Christoph Wurst <christoph@winzerhof-wurst.at>